Box Station CHANGES (Formerly Boxstation Tweaks) (#169)
* Update BoxStation_Lumos.dmm

* Fixes some crud I missed

Replaced the stools in Chapel with pews, more or less just to test the waters with players. Also added a spare bottle of holy water and some candles to Chapel office, Bartender closet, and the Private Study. Personally candles were great for RP and religous rituals, dunno why they all disappeared.

* Big big Changus big changus big changus

* More stuff

* Finalized changes

* Maint and nearstation changes

* Update BoxStation_Lumos.dmm

* Sweep of fixes and changes

Mostly in Maint, simplifies a section of pipes and such. Should be done.

* Single area update

Big shit

* Robotics surgery rework

* Update BoxStation_Lumos.dmm

* switches morgue and Starboard storage

* Pipe smoothing

* Update BoxStation_Lumos.dmm

* Update BoxStation_Lumos.dmm

* Update BoxStation_Lumos.dmm

* Update BoxStation_Lumos.dmm

* removes the stray gamera